+++ src/lib.rs
@@ -259,7 +259,7 @@
 //! printcln!("{}{'_}",gradient("BonnyAD9", (250, 50, 170), (180, 50, 240)));
 //! ```
 
-#![cfg_attr(docsrs, feature(doc_auto_cfg))]
+#![cfg_attr(docsrs, feature(doc_cfg))]
 
 pub use termal_core::*;
 pub use termal_proc as proc;