+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## future
+### Fixes
+- Properly use cache in `TermText::strip_control` and
+  `TermText::to_string_cache`
+
 ## v4.1.1
 ### Documentation
 - `raw::WaitForIn`